+++ csvtk/cmd/add-header.go
@@ -88,6 +88,13 @@ var addHeaderCmd = &cobra.Command{
 					checkError(writer.Write(record))
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 		writer.Flush()
 		checkError(writer.Error())