+++ csvtk/cmd/collapse.go
@@ -270,6 +270,13 @@ var collapseCmd = &cobra.Command{
 
 		writer.Flush()
 		checkError(writer.Error())
+
+		if config.IgnoreEmptyRow {
+			log.Warningf("%d empty rows ignored", csvReader.NumEmptyRows)
+		}
+		if config.IgnoreIllegalRow {
+			log.Warningf("%d illegal rows ignored", csvReader.NumIllegalRows)
+		}
 	},
 }