+++ csvtk/cmd/csv2json.go
@@ -271,6 +271,13 @@ var csv2jsonCmd = &cobra.Command{
 		} else {
 			outfh.WriteString("]\n")
 		}
+
+		if config.IgnoreEmptyRow {
+			log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+		}
+		if config.IgnoreIllegalRow {
+			log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+		}
 	},
 }