+++ csvtk/cmd/csv2tab.go
@@ -65,6 +65,13 @@ var csv2tabCmd = &cobra.Command{
 					checkError(writer.Write(record))
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 		writer.Flush()
 		checkError(writer.Error())