+++ csvtk/cmd/cut.go
@@ -232,6 +232,13 @@ var cutCmd = &cobra.Command{
 
 		writer.Flush()
 		checkError(writer.Error())
+
+		if config.IgnoreEmptyRow {
+			log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+		}
+		if config.IgnoreIllegalRow {
+			log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+		}
 	},
 }