+++ csvtk/cmd/dim.go
@@ -84,6 +84,13 @@ var dimCmd = &cobra.Command{
 				humanize.Comma(int64(numCols)),
 				humanize.Comma(int64(numRows)))
 
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
+
 		}
 		outfh.Write(tbl.Bytes())
 	},