+++ csvtk/cmd/filter2.go
@@ -301,6 +301,13 @@ Supported operators and types:
 					checkError(writer.Write(record))
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 		writer.Flush()
 		checkError(writer.Error())