+++ csvtk/cmd/gather.go
@@ -254,6 +254,13 @@ var gatherCmd = &cobra.Command{
 
 		writer.Flush()
 		checkError(writer.Error())
+
+		if config.IgnoreEmptyRow {
+			log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+		}
+		if config.IgnoreIllegalRow {
+			log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+		}
 	},
 }