+++ csvtk/cmd/grep.go
@@ -389,6 +389,13 @@ var grepCmd = &cobra.Command{
 					checkError(writer.Write(record))
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 		writer.Flush()
 		checkError(writer.Error())