+++ csvtk/cmd/header.go
@@ -66,6 +66,13 @@ var headersCmd = &cobra.Command{
 					break LOOP
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 	},
 }