+++ csvtk/cmd/mutate.go
@@ -237,6 +237,13 @@ var mutateCmd = &cobra.Command{
 					checkError(writer.Write(record2))
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 		writer.Flush()
 		checkError(writer.Error())