+++ csvtk/cmd/pretty.go
@@ -60,7 +60,7 @@ Attention:
 
 		file := files[0]
 		var colnames []string
-		headerRow, data := readCSV(config, file)
+		headerRow, data, csvReader := readCSV(config, file)
 
 		if len(headerRow) > 0 {
 			colnames = headerRow
@@ -99,6 +99,12 @@ Attention:
 			outfh.Write(tbl.Bytes())
 		}
 
+		if config.IgnoreEmptyRow {
+			log.Warningf("%d empty rows ignored", csvReader.NumEmptyRows)
+		}
+		if config.IgnoreIllegalRow {
+			log.Warningf("%d illegal rows ignored", csvReader.NumIllegalRows)
+		}
 	},
 }