+++ csvtk/cmd/transpose.go
@@ -86,6 +86,13 @@ var transposeCmd = &cobra.Command{
 					}
 				}
 			}
+
+			if config.IgnoreEmptyRow {
+				log.Warningf("file '%s': %d empty rows ignored", file, csvReader.NumEmptyRows)
+			}
+			if config.IgnoreIllegalRow {
+				log.Warningf("file '%s': %d illegal rows ignored", file, csvReader.NumIllegalRows)
+			}
 		}
 
 		writer := csv.NewWriter(outfh)